+++ src/mono/browser/runtime/dotnet.d.ts
@@ -126,7 +126,7 @@ interface DotnetHostBuilder {
      */
     create(): Promise<RuntimeAPI>;
     /**
-     * @deprecated use runMainAndExit() or runMain() instead.
+     * @deprecated use runMain() or runMainAndExit() instead.
      */
     run(): Promise<number>;
     /**
