+++ src/native/corehost/browserhost/host/cross-linked.ts
@@ -9,5 +9,4 @@ declare global {
     export function _BrowserHost_InitializeCoreCLR(): number;
     export function _BrowserHost_ExecuteAssembly(mainAssemblyNamePtr: number, argsLength: number, argsPtr: number): number;
     export function _wasm_load_icu_data(dataPtr: VoidPtr): number;
-    export function ExitStatus(exitCode: number): number;
 }
