+++ src/native/corehost/browserhost/host/index.ts
@@ -5,7 +5,7 @@ import type { InternalExchange, BrowserHostExports, RuntimeAPI, BrowserHostExpor
 import { InternalExchangeIndex } from "./types";
 import { } from "./cross-linked"; // ensure ambient symbols are declared
 
-import { runMain, runMainAndExit, registerDllBytes, installVfsFile, loadIcuData, getExitStatus, initializeCoreCLR } from "./host";
+import { runMain, runMainAndExit, registerDllBytes, installVfsFile, loadIcuData, initializeCoreCLR } from "./host";
 
 export function dotnetInitializeModule(internals: InternalExchange): void {
     if (!Array.isArray(internals)) throw new Error("Expected internals to be an array");
@@ -21,7 +21,6 @@ export function dotnetInitializeModule(internals: InternalExchange): void {
         registerDllBytes,
         installVfsFile,
         loadIcuData,
-        getExitStatus,
         initializeCoreCLR,
     });
     dotnetUpdateInternals(internals, dotnetUpdateInternalsSubscriber);
@@ -31,7 +30,6 @@ export function dotnetInitializeModule(internals: InternalExchange): void {
             map.registerDllBytes,
             map.installVfsFile,
             map.loadIcuData,
-            map.getExitStatus,
             map.initializeCoreCLR,
         ];
     }
