+++ src/native/libs/Common/JavaScript/types/exchange.ts
@@ -1,15 +1,18 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-import type { installVfsFile, registerDllBytes, loadIcuData, getExitStatus, initializeCoreCLR } from "../../../../corehost/browserhost/host/host";
 import type { check, error, info, warn, debug } from "../../../../corehost/browserhost/loader/logging";
-import type { createPromiseCompletionSource, getPromiseCompletionSource, isControllablePromise } from "../../../../corehost/browserhost/loader/promise-completion-source";
 import type { resolveRunMainPromise, rejectRunMainPromise, getRunMainPromise, abortStartup } from "../../../../corehost/browserhost/loader/run";
+import type { addOnExitListener, isExited, isRuntimeRunning, quitNow } from "../../../../corehost/browserhost/loader/exit";
+
+import type { installVfsFile, registerDllBytes, loadIcuData, initializeCoreCLR } from "../../../../corehost/browserhost/host/host";
+import type { createPromiseCompletionSource, getPromiseCompletionSource, isControllablePromise } from "../../../../corehost/browserhost/loader/promise-completion-source";
+
 import type { isSharedArrayBuffer, zeroRegion } from "../../../System.Native.Browser/utils/memory";
 import type { stringToUTF16, stringToUTF16Ptr, stringToUTF8Ptr, utf16ToString } from "../../../System.Native.Browser/utils/strings";
+import type { abortPosix, abortTimers, getExitStatus } from "../../../System.Native.Browser/utils/host";
+
 import type { symbolicateStackTrace } from "../../../System.Native.Browser/diagnostics/symbolicate";
-import type { addOnExitListener, isExited, isRuntimeRunning, quitNow } from "../../../../corehost/browserhost/loader/exit";
-import type { abortPosix, abortTimers } from "../../../System.Native.Browser/utils/host";
 
 export type RuntimeExports = {
 }
@@ -65,7 +68,6 @@ export type BrowserHostExports = {
     registerDllBytes: typeof registerDllBytes
     installVfsFile: typeof installVfsFile
     loadIcuData: typeof loadIcuData
-    getExitStatus: typeof getExitStatus
     initializeCoreCLR: typeof initializeCoreCLR
 }
 
@@ -73,7 +75,6 @@ export type BrowserHostExportsTable = [
     typeof registerDllBytes,
     typeof installVfsFile,
     typeof loadIcuData,
-    typeof getExitStatus,
     typeof initializeCoreCLR,
 ]
 
@@ -98,6 +99,7 @@ export type BrowserUtilsExports = {
     isSharedArrayBuffer: typeof isSharedArrayBuffer
     abortTimers: typeof abortTimers,
     abortPosix: typeof abortPosix,
+    getExitStatus: typeof getExitStatus,
 }
 
 export type BrowserUtilsExportsTable = [
@@ -109,6 +111,7 @@ export type BrowserUtilsExportsTable = [
     typeof isSharedArrayBuffer,
     typeof abortTimers,
     typeof abortPosix,
+    typeof getExitStatus,
 ]
 
 export type DiagnosticsExportsTable = [
